Split intersect.test.js into intersectGeom2.test.js and intersectGeom3.test.js
